feat: consolidate toa single task executor
