--- cuenv-proposal/ci.cue
@@ -1,10 +1,0 @@
-package env0
-
-import p0 "github.com/rawkode/blah/p0:env"
-
-import p1 "github.com/rawkode/blah/p1:env"
-
-envs: [
-	p0,
-	p1,
-]
