--- cuenv-proposal/cue.mod/module.cue
@@ -1,4 +1,0 @@
-module: "github.com/rawkode/blah"
-language: {
-	version: "v0.13.2"
-}
