--- test_tmp/cue.mod/module.cue
@@ -1,1 +1,0 @@
-module: "test.example/monorepo"
