--- test_tmp/env.cue
@@ -1,3 +1,0 @@
-package env
-
-env: {ROOT: "true"}
