--- test_tmp/packages/api/env.cue
@@ -1,10 +1,0 @@
-package env
-
-env: {API: "true"}
-tasks: {
-	"build": {
-		command:     "echo api > build.txt"
-		description: "Build API"
-		dependencies: ["packages:web:build"]
-	}
-}
