--- test_tmp/packages/web/env.cue
@@ -1,9 +1,0 @@
-package env
-
-env: {WEB: "true"}
-tasks: {
-	"build": {
-		command:     "echo web > build.txt"
-		description: "Build web"
-	}
-}
